requirements.txt
Django==4.2.0
python-decouple==3.8
djangorestframework==3.16.0
uvicorn==0.34.2
channels==4.1.0
channels-redis==4.2.0
whitenoise==6.9.0
django-redis==5.4.0
psycopg2==2.9.4